MissleSystem -> StackWeapon 이름 변경 외 2
- UI 업데이트 기능을 PlayerShip으로 이전
- 스택이 소수점 단위로 쌓이도록 수정

+++ Assets/Scripts/UIs/CombatUiManager.cs
@@ -52,7 +52,7 @@ public class CombatUiManager : MonoSingleton<CombatUiManager>
         healthUI.InitUI(maxHealth / 100, (float)currHealth / 100);        
     }
 
-    public void SetMissleUI(int currMissle, int maxMissle)
+    public void SetMissleUI(float currMissle, int maxMissle)
     {
         missleUI.InitUI(maxMissle, currMissle);
     }
